Cloud Solutions/Mixit/backend/requirements.txt
Flask>=3.0.3
werkzeug>=2.2
flask-session~=0.8.0
requests>=2
markupsafe==2.1.1
git+https://github.com/azure-samples/ms-identity-python-samples-common
pyopenssl==24.0.0
flask-cors==4.0.0
identity>=0.5.1,<0.6
python-dotenv<0.22
flask_sqlalchemy== 3.0.2
datetime
pyodbc
python-pptx==0.6.22
python-docx==1.1.2
spire.presentation==9.3.0
spire.doc==12.4.0
pymupdf==1.24.5